Adds error checking for incomplete arrays.
When processing a stream of bytes (e.g., network buffer) you may not have all of the bytes of a character yet.  In this case, the code would attempt to read past the end of the array and throw an exception that wasn't as useful as it could be.  This just adds some basic length checks to the function to ensure reading off the end of the array doesn't happen.